pi/requirements.txt
colorama==0.3.9
cursor==1.2.0
enum34==1.1.6
log-symbols==0.0.12
paho-mqtt==1.4.0
pkg-resources==0.0.0
six==1.11.0
